+++ src/main/java/br/com/storemanager/persistence/ProductRepository.java
@@ -1,7 +1,7 @@
 package br.com.storemanager.persistence;
 
-import br.com.storemanager.model.Product;
+import br.com.storemanager.model.product.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 }
